backend/utils/gemini_adapter/__init__.py
"""
OpenAI-Gemini compatibility adapter that allows using Google Gemini through an OpenAI-like interface.
This adapter simulates the interface that would be used in the actual implementation.
"""
import os
from typing import List, Dict, Any, Optional
import google.generativeai as genai
from openai import OpenAI


class GeminiAdapter:
    """
    Adapter class that provides an OpenAI-like interface for Google Gemini.
    This allows using the same tools and patterns with either provider.
    """
    def __init__(self):
        # Initialize Google Gemini
        api_key = os.getenv("GOOGLE_GEMINI_API_KEY")
        if not api_key:
            raise ValueError("GOOGLE_GEMINI_API_KEY environment variable is required")

        genai.configure(api_key=api_key)
        self.model = genai.GenerativeModel('gemini-pro')

        # Also keep an OpenAI client reference for compatibility
        self.openai_client = None  # We won't actually initialize this in this mock

    def chat_completions_create(self, model: str, messages: List[Dict[str, str]], tools: Optional[List[Dict[str, Any]]] = None, **kwargs):
        """
        Simulate the OpenAI chat completions API interface for Gemini.

        Args:
            model: The model identifier (ignored in this adapter)
            messages: List of message dictionaries with role and content
            tools: Optional list of tools available to the model
            **kwargs: Additional parameters

        Returns:
            Mock response that mimics OpenAI API format
        """
        # In a real implementation, this would call the actual Gemini API
        # with proper message conversion and tool integration

        # For this implementation, we'll return a mock response
        # since we're focusing on the backend architecture
        from types import SimpleNamespace
        import json

        # This is a simplified version for the purpose of demonstrating the structure
        # In a real implementation, we'd call the Gemini API and convert the response
        last_message = messages[-1]['content'] if messages else ""

        # Mock response structure to match OpenAI format
        response = SimpleNamespace(
            choices=[
                SimpleNamespace(
                    message=SimpleNamespace(
                        role="assistant",
                        content=f"I received your message: '{last_message}'. The AI processing would happen here.",
                        tool_calls=[]  # In real implementation, this would contain actual tool calls
                    ),
                    finish_reason="stop"
                )
            ],
            usage={
                "prompt_tokens": 10,
                "completion_tokens": 20,
                "total_tokens": 30
            }
        )

        return response


# Global instance for compatibility
_adapter = None


def get_gemini_adapter():
    """
    Get or create the singleton Gemini adapter instance.
    """
    global _adapter
    if _adapter is None:
        _adapter = GeminiAdapter()
    return _adapter